fix doc: it doesn't support LongTensor, converting to float

+++ torch/nn/modules/upsampling.py
@@ -50,7 +50,7 @@ class Upsample(Module):
 
     Examples::
 
-        >>> input = torch.arange(1, 5).view(1, 1, 2, 2)
+        >>> input = torch.arange(1, 5).view(1, 1, 2, 2).float()
         >>> input
         tensor([[[[ 1.,  2.],
                   [ 3.,  4.]]]])
